requirements.txt
Django==1.6.2
PySocks==1.5.0
celery==3.1.9
